display a space between zones
Because  we draw  hexagons, some  drawed  lines are  not horizontal  nor
vertical, so they are drawed using the Bresenham's line algorithm and it
it is  harder to  assert that  2 zones won't  have any  collapsing pixel
(which is  visible by  the user).  In conclusion  I prefers  display the
border, and make the result homogeneous and predictible.